Third parties scripts refactoring
