+++ packages/tasit-contracts/test/LANDRegistry.js
@@ -1,6 +1,6 @@
-const abi = require("../decentraland/land/build/contracts/LANDRegistry.json")
+const abi = require("../3rd-parties/decentraland/land/build/contracts/LANDRegistry.json")
   .abi;
-const localAddresses = require("../decentraland/addresses").local;
+const localAddresses = require("../3rd-parties/decentraland/addresses").local;
 const { LANDRegistry: address } = localAddresses;
 
 // Note:
